Refactor parameters
