+++ src/main/java/com/powsybl/openloadflow/OpenLoadFlowProvider.java
@@ -8,12 +8,9 @@ package com.powsybl.openloadflow;
 
 import com.google.auto.service.AutoService;
 import com.google.common.base.Stopwatch;
-import com.powsybl.commons.PowsyblException;
 import com.powsybl.commons.reporter.Reporter;
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.iidm.network.Network;
-import com.powsybl.iidm.network.PhaseTapChanger;
-import com.powsybl.iidm.network.TwoWindingsTransformer;
 import com.powsybl.iidm.network.extensions.SlackTerminal;
 import com.powsybl.loadflow.LoadFlowParameters;
 import com.powsybl.loadflow.LoadFlowProvider;
@@ -22,36 +19,27 @@ import com.powsybl.loadflow.LoadFlowResultImpl;
 import com.powsybl.loadflow.resultscompletion.z0flows.Z0FlowsCompletion;
 import com.powsybl.math.matrix.MatrixFactory;
 import com.powsybl.math.matrix.SparseMatrixFactory;
-import com.powsybl.openloadflow.ac.DefaultOuterLoopConfig;
-import com.powsybl.openloadflow.ac.equations.AcEquationSystemCreationParameters;
-import com.powsybl.openloadflow.ac.nr.DefaultNewtonRaphsonStoppingCriteria;
-import com.powsybl.openloadflow.ac.nr.NewtonRaphsonParameters;
 import com.powsybl.openloadflow.ac.nr.NewtonRaphsonStatus;
-import com.powsybl.openloadflow.ac.outerloop.*;
+import com.powsybl.openloadflow.ac.outerloop.AcLoadFlowParameters;
+import com.powsybl.openloadflow.ac.outerloop.AcLoadFlowResult;
+import com.powsybl.openloadflow.ac.outerloop.AcloadFlowEngine;
+import com.powsybl.openloadflow.ac.outerloop.OuterLoop;
 import com.powsybl.openloadflow.dc.DcLoadFlowEngine;
-import com.powsybl.openloadflow.dc.DcLoadFlowParameters;
 import com.powsybl.openloadflow.dc.DcLoadFlowResult;
-import com.powsybl.openloadflow.dc.DcValueVoltageInitializer;
-import com.powsybl.openloadflow.dc.equations.DcEquationSystemCreationParameters;
-import com.powsybl.openloadflow.ac.VoltageMagnitudeInitializer;
-import com.powsybl.openloadflow.network.LfNetworkParameters;
-import com.powsybl.openloadflow.network.NetworkSlackBusSelector;
 import com.powsybl.openloadflow.network.PerUnit;
-import com.powsybl.openloadflow.network.SlackBusSelector;
 import com.powsybl.openloadflow.network.impl.LfNetworkLoaderImpl;
 import com.powsybl.openloadflow.network.impl.Networks;
-import com.powsybl.openloadflow.network.util.PreviousValueVoltageInitializer;
-import com.powsybl.openloadflow.network.util.UniformValueVoltageInitializer;
-import com.powsybl.openloadflow.network.util.VoltageInitializer;
 import com.powsybl.openloadflow.util.Markers;
 import com.powsybl.openloadflow.util.PowsyblOpenLoadFlowVersion;
 import com.powsybl.tools.PowsyblCoreVersion;
 import net.jafama.FastMath;
-import org.apache.commons.compress.utils.Lists;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
@@ -94,146 +82,16 @@ public class OpenLoadFlowProvider implements LoadFlowProvider {
         return new PowsyblCoreVersion().getMavenProjectVersion();
     }
 
-    static VoltageInitializer getVoltageInitializer(LoadFlowParameters parameters, LfNetworkParameters networkParameters, MatrixFactory matrixFactory, Reporter reporter) {
-        switch (parameters.getVoltageInitMode()) {
-            case UNIFORM_VALUES:
-                return new UniformValueVoltageInitializer();
-            case PREVIOUS_VALUES:
-                return new PreviousValueVoltageInitializer();
-            case DC_VALUES:
-                return new DcValueVoltageInitializer(networkParameters, parameters.isDistributedSlack(), parameters.getBalanceType(), parameters.isDcUseTransformerRatio(), matrixFactory, reporter);
-            default:
-                throw new UnsupportedOperationException("Unsupported voltage init mode: " + parameters.getVoltageInitMode());
-        }
-    }
+    private LoadFlowResult runAc(Network network, LoadFlowParameters parameters, Reporter reporter) {
+        OpenLoadFlowParameters parametersExt = OpenLoadFlowParameters.get(parameters);
+        OpenLoadFlowParameters.logAc(parameters, parametersExt);
 
-    static VoltageInitializer getExtendedVoltageInitializer(LoadFlowParameters parameters, OpenLoadFlowParameters parametersExt,
-                                                            LfNetworkParameters networkParameters, MatrixFactory matrixFactory, Reporter reporter) {
-        switch (parametersExt.getVoltageInitModeOverride()) {
-            case NONE:
-                return getVoltageInitializer(parameters, networkParameters, matrixFactory, reporter);
+        AcLoadFlowParameters acParameters = OpenLoadFlowParameters.createAcParameters(network, parameters, parametersExt, matrixFactory, reporter);
 
-            case VOLTAGE_MAGNITUDE:
-                return new VoltageMagnitudeInitializer(matrixFactory);
-
-            case FULL_VOLTAGE:
-                return new FullVoltageInitializer(new VoltageMagnitudeInitializer(matrixFactory),
-                                                  new DcValueVoltageInitializer(networkParameters,
-                                                                                parameters.isDistributedSlack(),
-                                                                                parameters.getBalanceType(),
-                                                                                parameters.isDcUseTransformerRatio(),
-                                                                                matrixFactory,
-                                                                                reporter));
-
-            default:
-                throw new PowsyblException("Unknown voltage init mode override: " + parametersExt.getVoltageInitModeOverride());
-        }
-    }
-
-    static SlackBusSelector getSlackBusSelector(Network network, LoadFlowParameters parameters, OpenLoadFlowParameters parametersExt) {
-        SlackBusSelector slackBusSelector = SlackBusSelector.fromMode(parametersExt.getSlackBusSelectionMode(), parametersExt.getSlackBusesIds());
-        return parameters.isReadSlackBus() ? new NetworkSlackBusSelector(network, slackBusSelector)
-                                           : slackBusSelector;
-    }
-
-    private static OuterLoopConfig findOuterLoopConfig() {
-        OuterLoopConfig outerLoopConfig;
-        List<OuterLoopConfig> outerLoopConfigs = Lists.newArrayList(ServiceLoader.load(OuterLoopConfig.class, OuterLoopConfig.class.getClassLoader()).iterator());
-        if (outerLoopConfigs.isEmpty()) {
-            outerLoopConfig = new DefaultOuterLoopConfig();
-        } else {
-            if (outerLoopConfigs.size() > 1) {
-                throw new PowsyblException("Only one outer loop config is expected on class path");
-            }
-            outerLoopConfig = outerLoopConfigs.get(0);
-        }
-        return outerLoopConfig;
-    }
-
-    public static AcLoadFlowParameters createAcParameters(Network network, MatrixFactory matrixFactory, LoadFlowParameters parameters,
-                                                          OpenLoadFlowParameters parametersExt, boolean breakers, Reporter reporter) {
-        Set<String> branchesWithCurrent = null;
-        if (parameters.isPhaseShifterRegulationOn()) {
-            branchesWithCurrent = network.getTwoWindingsTransformerStream()
-                    .filter(twt -> twt.getPhaseTapChanger() != null && twt.getPhaseTapChanger().getRegulationMode() == PhaseTapChanger.RegulationMode.CURRENT_LIMITER)
-                    .map(TwoWindingsTransformer::getId)
-                    .collect(Collectors.toSet());
-        }
-        return createAcParameters(network, matrixFactory, parameters, parametersExt, breakers, false, branchesWithCurrent, reporter);
-    }
-
-    static LfNetworkParameters getNetworkParameters(LoadFlowParameters parameters, OpenLoadFlowParameters parametersExt,
-                                                    SlackBusSelector slackBusSelector, boolean breakers) {
-        return new LfNetworkParameters(slackBusSelector,
-                                       parametersExt.hasVoltageRemoteControl(),
-                                       parametersExt.getLowImpedanceBranchMode() == OpenLoadFlowParameters.LowImpedanceBranchMode.REPLACE_BY_MIN_IMPEDANCE_LINE,
-                                       parameters.isTwtSplitShuntAdmittance(),
-                                       breakers,
-                                       parametersExt.getPlausibleActivePowerLimit(),
-                                       parametersExt.isAddRatioToLinesWithDifferentNominalVoltageAtBothEnds(),
-                                       parameters.getConnectedComponentMode() == LoadFlowParameters.ConnectedComponentMode.MAIN,
-                                       parameters.getCountriesToBalance(),
-                                       parameters.isDistributedSlack() && parameters.getBalanceType() == LoadFlowParameters.BalanceType.PROPORTIONAL_TO_CONFORM_LOAD,
-                                       parameters.isPhaseShifterRegulationOn(),
-                                       parameters.isTransformerVoltageControlOn(),
-                                       parametersExt.isVoltagePerReactivePowerControl(),
-                                       parametersExt.hasReactivePowerRemoteControl(),
-                                       parameters.isDc(),
-                                       parameters.isSimulShunt(),
-                                       !parameters.isNoGeneratorReactiveLimits());
-    }
-
-    public static AcLoadFlowParameters createAcParameters(Network network, MatrixFactory matrixFactory, LoadFlowParameters parameters,
-                                                          OpenLoadFlowParameters parametersExt, boolean breakers, boolean forceA1Var,
-                                                          Set<String> branchesWithCurrent, Reporter reporter) {
-
-        SlackBusSelector slackBusSelector = getSlackBusSelector(network, parameters, parametersExt);
-
-        LOGGER.info("Slack bus selector: {}", slackBusSelector.getClass().getSimpleName());
-        LOGGER.info("Voltage initialization mode: {}", parameters.getVoltageInitMode());
-        LOGGER.info("Voltage initialization mode override: {}", parametersExt.getVoltageInitModeOverride());
-        LOGGER.info("Distributed slack: {}", parameters.isDistributedSlack());
-        LOGGER.info("Balance type: {}", parameters.getBalanceType());
-        LOGGER.info("Reactive limits: {}", !parameters.isNoGeneratorReactiveLimits());
-        LOGGER.info("Voltage remote control: {}", parametersExt.hasVoltageRemoteControl());
-        LOGGER.info("Phase control: {}", parameters.isPhaseShifterRegulationOn());
-        LOGGER.info("Split shunt admittance: {}", parameters.isTwtSplitShuntAdmittance());
-        LOGGER.info("Direct current: {}", parameters.isDc());
-        LOGGER.info("Transformer voltage control: {}", parameters.isTransformerVoltageControlOn());
-        LOGGER.info("Load power factor constant: {}", parametersExt.isLoadPowerFactorConstant());
-        LOGGER.info("Plausible active power limit: {}", parametersExt.getPlausibleActivePowerLimit());
-        LOGGER.info("Add ratio to lines with different nominal voltage at both ends: {}", parametersExt.isAddRatioToLinesWithDifferentNominalVoltageAtBothEnds());
-        LOGGER.info("Slack bus Pmax mismatch: {}", parametersExt.getSlackBusPMaxMismatch());
-        LOGGER.info("Connected component mode: {}", parameters.getConnectedComponentMode());
-        LOGGER.info("Voltage per reactive power control: {}", parametersExt.isVoltagePerReactivePowerControl());
-        LOGGER.info("Reactive Power Remote control: {}", parametersExt.hasReactivePowerRemoteControl());
-        LOGGER.info("Shunt voltage control: {}", parameters.isSimulShunt());
-
-        OuterLoopConfig outerLoopConfig = findOuterLoopConfig();
-        List<OuterLoop> outerLoops = outerLoopConfig.configure(parameters, parametersExt);
         if (LOGGER.isInfoEnabled()) {
-            LOGGER.info("Outer loops: {}", outerLoops.stream().map(OuterLoop::getType).collect(Collectors.toList()));
+            LOGGER.info("Outer loops: {}", acParameters.getOuterLoops().stream().map(OuterLoop::getType).collect(Collectors.toList()));
         }
 
-        var networkParameters = getNetworkParameters(parameters, parametersExt, slackBusSelector, breakers);
-
-        var equationSystemCreationParameters = new AcEquationSystemCreationParameters(forceA1Var, branchesWithCurrent);
-
-        VoltageInitializer voltageInitializer = getExtendedVoltageInitializer(parameters, parametersExt, networkParameters, matrixFactory, reporter);
-
-        var newtonRaphsonParameters = new NewtonRaphsonParameters()
-                .setVoltageInitializer(voltageInitializer)
-                .setStoppingCriteria(new DefaultNewtonRaphsonStoppingCriteria(parametersExt.getNewtonRaphsonConvEpsPerEq()))
-                .setMaxIteration(parametersExt.getMaxIteration());
-
-        return new AcLoadFlowParameters(networkParameters,
-                                        equationSystemCreationParameters,
-                                        newtonRaphsonParameters,
-                                        outerLoops, matrixFactory);
-    }
-
-    private LoadFlowResult runAc(Network network, LoadFlowParameters parameters, OpenLoadFlowParameters parametersExt, Reporter reporter) {
-        AcLoadFlowParameters acParameters = createAcParameters(network, matrixFactory, parameters, parametersExt, false, reporter);
         List<AcLoadFlowResult> results = AcloadFlowEngine.run(network, new LfNetworkLoaderImpl(), acParameters, reporter);
 
         Networks.resetState(network);
@@ -294,53 +152,11 @@ public class OpenLoadFlowProvider implements LoadFlowProvider {
         return new LoadFlowResultImpl(ok, Collections.emptyMap(), null, componentResults);
     }
 
-    static DcLoadFlowParameters createDcParameters(Network network, MatrixFactory matrixFactory,
-                                                   LoadFlowParameters parameters, OpenLoadFlowParameters parametersExt,
-                                                   boolean forcePhaseControlOffAndAddAngle1Var) {
-        SlackBusSelector slackBusSelector = getSlackBusSelector(network, parameters, parametersExt);
+    private LoadFlowResult runDc(Network network, LoadFlowParameters parameters, Reporter reporter) {
+        OpenLoadFlowParameters parametersExt = OpenLoadFlowParameters.get(parameters);
+        OpenLoadFlowParameters.logDc(parameters, parametersExt);
 
-        LOGGER.info("Slack bus selector: {}", slackBusSelector.getClass().getSimpleName());
-        LOGGER.info("Use transformer ratio: {}", parameters.isDcUseTransformerRatio());
-        LOGGER.info("Distributed slack: {}", parameters.isDistributedSlack());
-        LOGGER.info("Balance type: {}", parameters.getBalanceType());
-        LOGGER.info("Plausible active power limit: {}", parametersExt.getPlausibleActivePowerLimit());
-        LOGGER.info("Add ratio to lines with different nominal voltage at both ends: {}", parametersExt.isAddRatioToLinesWithDifferentNominalVoltageAtBothEnds());
-        LOGGER.info("Connected component mode: {}", parameters.getConnectedComponentMode());
-
-        var networkParameters = new LfNetworkParameters(slackBusSelector,
-                                                        false,
-                                                        false,
-                                                        false,
-                                                        false,
-                                                        parametersExt.getPlausibleActivePowerLimit(),
-                                                        false,
-                                                        parameters.getConnectedComponentMode() == LoadFlowParameters.ConnectedComponentMode.MAIN,
-                                                        parameters.getCountriesToBalance(),
-                                                        parameters.isDistributedSlack() && parameters.getBalanceType() == LoadFlowParameters.BalanceType.PROPORTIONAL_TO_CONFORM_LOAD,
-                                                        false,
-                                                        false,
-                                                        false,
-                                                        false,
-                                                        true,
-                                                        false,
-                                                        false);
-
-        var equationSystemCreationParameters = new DcEquationSystemCreationParameters(true,
-                                                                                      false,
-                                                                                      forcePhaseControlOffAndAddAngle1Var,
-                                                                                      parameters.isDcUseTransformerRatio());
-
-        return new DcLoadFlowParameters(networkParameters,
-                                        equationSystemCreationParameters,
-                                        matrixFactory,
-                                        parameters.isDistributedSlack(),
-                                        parameters.getBalanceType(),
-                                        true);
-    }
-
-    private LoadFlowResult runDc(Network network, LoadFlowParameters parameters, OpenLoadFlowParameters parametersExt, Reporter reporter) {
-
-        var dcParameters = createDcParameters(network, matrixFactory, parameters, parametersExt, forcePhaseControlOffAndAddAngle1Var);
+        var dcParameters = OpenLoadFlowParameters.createDcParameters(network, parameters, parametersExt, matrixFactory, forcePhaseControlOffAndAddAngle1Var);
 
         List<DcLoadFlowResult> results = new DcLoadFlowEngine(network, new LfNetworkLoaderImpl(), dcParameters, reporter)
                 .run(reporter);
@@ -387,8 +203,6 @@ public class OpenLoadFlowProvider implements LoadFlowProvider {
 
         LOGGER.info("Version: {}", new PowsyblOpenLoadFlowVersion());
 
-        OpenLoadFlowParameters parametersExt = OpenLoadFlowParameters.get(parameters);
-
         Reporter lfReporter = reporter.createSubReporter("loadFlow", "Load flow on network ${networkId}",
             "networkId", network.getId());
 
@@ -398,8 +212,8 @@ public class OpenLoadFlowProvider implements LoadFlowProvider {
 
             Stopwatch stopwatch = Stopwatch.createStarted();
 
-            LoadFlowResult result = parameters.isDc() ? runDc(network, parameters, parametersExt, lfReporter)
-                                                      : runAc(network, parameters, parametersExt, lfReporter);
+            LoadFlowResult result = parameters.isDc() ? runDc(network, parameters, lfReporter)
+                                                      : runAc(network, parameters, lfReporter);
 
             stopwatch.stop();
             LOGGER.info(Markers.PERFORMANCE_MARKER, "Load flow ran in {} ms", stopwatch.elapsed(TimeUnit.MILLISECONDS));
