+++ src/test/java/com/powsybl/openloadflow/network/impl/LfSwitchTest.java
@@ -12,7 +12,6 @@ import com.powsybl.iidm.network.Network;
 import com.powsybl.loadflow.LoadFlowParameters;
 import com.powsybl.math.matrix.DenseMatrixFactory;
 import com.powsybl.openloadflow.OpenLoadFlowParameters;
-import com.powsybl.openloadflow.OpenLoadFlowProvider;
 import com.powsybl.openloadflow.ac.equations.*;
 import com.powsybl.openloadflow.ac.outerloop.AcLoadFlowParameters;
 import com.powsybl.openloadflow.equations.EquationTerm;
@@ -44,8 +43,8 @@ class LfSwitchTest {
     @BeforeEach
     void setUp() {
         network = NodeBreakerNetworkFactory.create();
-        acLoadFlowParameters = OpenLoadFlowProvider.createAcParameters(network, new DenseMatrixFactory(), new LoadFlowParameters(),
-                new OpenLoadFlowParameters(), true, Reporter.NO_OP);
+        acLoadFlowParameters = OpenLoadFlowParameters.createAcParameters(network, new LoadFlowParameters(),
+                new OpenLoadFlowParameters(), new DenseMatrixFactory(), Reporter.NO_OP, true, false);
         List<LfNetwork> lfNetworks = Networks.load(network, acLoadFlowParameters.getNetworkParameters(), Reporter.NO_OP);
         assertEquals(1, lfNetworks.size());
         lfNetwork = lfNetworks.get(0);
